+++ lib/membrane/core/bin/pad_controller.ex
@@ -196,7 +196,9 @@ defmodule Membrane.Core.Bin.PadController do
   """
   @spec handle_unlink(Pad.ref_t(), Core.Bin.State.t()) :: Type.stateful_try_t(Core.Bin.State.t())
   def handle_unlink(pad_ref, state) do
-    with {:ok, state} <- maybe_handle_pad_removed(pad_ref, state) do
+    with {:ok, state} <- maybe_handle_pad_removed(pad_ref, state),
+         {:ok, endpoint} <- PadModel.get_data(state, pad_ref, :endpoint) do
+      Message.send(endpoint.pid, :handle_unlink, endpoint.pad_ref)
       PadModel.delete_data(state, pad_ref)
     end
   end