+++ test/membrane/element_test.exs
@@ -57,17 +57,17 @@ defmodule Membrane.ElementTest do
   end
 
   setup do
+    children = [
+      source: %Testing.Source{output: ['a', 'b', 'c']},
+      filter: %TestFilter{target: self()},
+      sink: Testing.Sink
+    ]
+
     {:ok, pipeline} =
-      Testing.Pipeline.start_link(%Testing.Pipeline.Options{
-        elements: [
-          source: %Testing.Source{output: ['a', 'b', 'c']},
-          filter: %TestFilter{target: self()},
-          sink: Testing.Sink
-        ]
-      })
+      Testing.Pipeline.start_link(links: Membrane.ParentSpec.link_linear(children))
 
     on_exit(fn ->
-      Membrane.Pipeline.stop_and_terminate(pipeline, blocking?: true)
+      Membrane.Pipeline.terminate(pipeline, blocking?: true)
     end)
 
     [pipeline: pipeline]
@@ -75,44 +75,36 @@ defmodule Membrane.ElementTest do
 
   describe "Start of stream" do
     test "causes handle_start_of_stream/3 to be called", %{pipeline: pipeline} do
-      Testing.Pipeline.play(pipeline)
       assert_pipeline_playback_changed(pipeline, _from, :playing)
 
       TestFilter.assert_callback_called(:handle_start_of_stream)
     end
 
     test "does not trigger calling callback handle_event/3", %{pipeline: pipeline} do
-      Testing.Pipeline.play(pipeline)
       assert_pipeline_playback_changed(pipeline, _from, :playing)
 
       TestFilter.refute_callback_called(:handle_event)
     end
 
     test "causes handle_element_start_of_stream/3 to be called in pipeline", %{pipeline: pipeline} do
-      Testing.Pipeline.play(pipeline)
-
       assert_start_of_stream(pipeline, :filter)
     end
   end
 
   describe "End of stream" do
     test "causes handle_end_of_stream/3 to be called", %{pipeline: pipeline} do
-      Testing.Pipeline.play(pipeline)
       assert_pipeline_playback_changed(pipeline, _from, :playing)
 
       TestFilter.assert_callback_called(:handle_end_of_stream)
     end
 
     test "does not trigger calling callback handle_event/3", %{pipeline: pipeline} do
-      Testing.Pipeline.play(pipeline)
       assert_pipeline_playback_changed(pipeline, _from, :playing)
 
       TestFilter.refute_callback_called(:handle_event)
     end
 
     test "causes handle_element_end_of_stream/3 to be called in pipeline", %{pipeline: pipeline} do
-      Testing.Pipeline.play(pipeline)
-
       assert_end_of_stream(pipeline, :filter)
     end
   end