+++ test/membrane/integration/demands_test.exs
@@ -17,8 +17,6 @@ defmodule Membrane.Integration.DemandsTest do
 
   defp test_pipeline(pid) do
     pattern_gen = fn i -> %Buffer{payload: <<i::16>> <> <<255>>} end
-    assert Pipeline.play(pid) == :ok
-
     assert_pipeline_playback_changed(pid, :prepared, :playing)
 
     demand = 500
@@ -34,18 +32,17 @@ defmodule Membrane.Integration.DemandsTest do
     demand..(2 * demand - 1)
     |> assert_buffers_received(pid)
 
-    assert Pipeline.stop_and_terminate(pid, blocking?: true) == :ok
+    assert Pipeline.terminate(pid, blocking?: true) == :ok
   end
 
   test "Regular pipeline with proper demands" do
-    assert {:ok, pid} =
-             Pipeline.start_link(%Pipeline.Options{
-               elements: [
-                 source: Source,
-                 filter: Filter,
-                 sink: %Sink{autodemand: false}
-               ]
-             })
+    children = [
+      source: Source,
+      filter: Filter,
+      sink: %Sink{autodemand: false}
+    ]
+
+    assert {:ok, pid} = Pipeline.start_link(links: Membrane.ParentSpec.link_linear(children))
 
     test_pipeline(pid)
   end
@@ -53,14 +50,13 @@ defmodule Membrane.Integration.DemandsTest do
   test "Pipeline with filter underestimating demand" do
     filter_demand_gen = fn _incoming_demand -> 2 end
 
-    assert {:ok, pid} =
-             Pipeline.start_link(%Pipeline.Options{
-               elements: [
-                 source: Source,
-                 filter: %Filter{demand_generator: filter_demand_gen},
-                 sink: %Sink{autodemand: false}
-               ]
-             })
+    children = [
+      source: Source,
+      filter: %Filter{demand_generator: filter_demand_gen},
+      sink: %Sink{autodemand: false}
+    ]
+
+    assert {:ok, pid} = Pipeline.start_link(links: Membrane.ParentSpec.link_linear(children))
 
     test_pipeline(pid)
   end
@@ -79,14 +75,13 @@ defmodule Membrane.Integration.DemandsTest do
       ~> {&1, cnt + 4}
     end
 
-    assert {:ok, pid} =
-             Pipeline.start_link(%Pipeline.Options{
-               elements: [
-                 source: %Source{output: {0, actions_gen}},
-                 filter: Filter,
-                 sink: %Sink{autodemand: false}
-               ]
-             })
+    children = [
+      source: %Source{output: {0, actions_gen}},
+      filter: Filter,
+      sink: %Sink{autodemand: false}
+    ]
+
+    assert {:ok, pid} = Pipeline.start_link(links: Membrane.ParentSpec.link_linear(children))
 
     test_pipeline(pid)
   end