+++ test/membrane/testing/pipeline_test.exs
@@ -11,13 +11,14 @@ defmodule Membrane.Testing.PipelineTest do
     def handle_init(_opts), do: {{:ok, spec: %Membrane.ParentSpec{}}, :state}
   end
 
-  describe "When initializing Testing Pipeline" do
-    test "generates links if only elements were provided" do
+  describe "Testing pipeline creation" do
+    test "works with :default implementation" do
       import ParentSpec
       elements = [elem: Elem, elem2: Elem]
       links = [link(:elem) |> to(:elem2)]
-      options = %Pipeline.Options{elements: elements}
-      assert {{:ok, spec: spec}, state} = Pipeline.handle_init(options)
+      options = %{module: :default, children: elements, links: links, test_process: nil}
+      assert {{:ok, spec: spec, playback: :playing}, state} = Pipeline.handle_init(options)
+
       assert state == %Pipeline.State{module: nil, test_process: nil}
 
       assert spec == %Membrane.ParentSpec{
@@ -26,12 +27,40 @@ defmodule Membrane.Testing.PipelineTest do
              }
     end
 
+    test "by default chooses :default implementation" do
+      import ParentSpec
+      elements = [elem: Elem, elem2: Elem]
+      links = [link(:elem) |> to(:elem2)]
+      options = %{module: :default, children: elements, links: links, test_process: nil}
+      assert {{:ok, spec: spec, playback: :playing}, state} = Pipeline.handle_init(options)
+      assert state == %Pipeline.State{module: nil, test_process: nil}
+
+      assert spec == %Membrane.ParentSpec{
+               links: links,
+               children: elements
+             }
+    end
+
+    test "works with custom module injected" do
+      options = %{module: MockPipeline, test_process: nil, custom_args: []}
+      assert {{:ok, spec: spec}, state} = Pipeline.handle_init(options)
+      assert spec == %Membrane.ParentSpec{}
+
+      assert state == %Pipeline.State{
+               custom_pipeline_state: :state,
+               module: MockPipeline,
+               test_process: nil
+             }
+    end
+  end
+
+  describe "When initializing Testing Pipeline" do
     test "uses prepared links if they were provided" do
       import ParentSpec
       elements = [elem: Elem, elem2: Elem]
       links = link(:elem) |> to(:elem2)
-      options = %Pipeline.Options{elements: elements, links: links}
-      assert {{:ok, spec: spec}, state} = Pipeline.handle_init(options)
+      options = %{module: :default, children: elements, links: links, test_process: nil}
+      assert {{:ok, spec: spec, playback: :playing}, state} = Pipeline.handle_init(options)
       assert state == %Pipeline.State{module: nil, test_process: nil}
 
       assert spec == %Membrane.ParentSpec{
@@ -53,16 +82,14 @@ defmodule Membrane.Testing.PipelineTest do
     end
   end
 
-  describe "When starting Testing Pipeline does" do
-    test "raises an error if a pipeline is started with both elements and module provided in options" do
-      assert_raise RuntimeError, ~r/override module and elements list/, fn ->
-        Pipeline.start(%Pipeline.Options{elements: [elem: Elem], module: Mod})
+  describe "When starting, Testing Pipeline" do
+    test "raises an error if not a module or non-existing module was passed" do
+      assert_raise RuntimeError, ~r/Not a module./, fn ->
+        Pipeline.start(module: [1, 2])
       end
-    end
 
-    test "raises an error if no means of generating spec are provided (no elements, no module)" do
-      assert_raise RuntimeError, ~r/You provided no information about pipeline contents./, fn ->
-        Pipeline.start(%Pipeline.Options{elements: nil, module: nil})
+      assert_raise RuntimeError, ~r/Unknown module./, fn ->
+        Pipeline.start(module: NotExistingModule)
       end
     end
   end